+++ wire-ios/Wire-iOS Tests/AppStateCalculatorTests.swift
@@ -127,6 +127,19 @@ final class AppStateCalculatorTests: XCTestCase {
         XCTAssertTrue(delegate.wasNotified)
     }
 
+    func testThatAppStateChanges_OnDidFailToLogin_CanNotRegisterMoreClients() {
+        // GIVEN
+        let error = NSError(code: ZMUserSessionErrorCode.canNotRegisterMoreClients, userInfo: nil)
+        sut.applicationDidBecomeActive()
+
+        // WHEN
+        sut.sessionManagerDidFailToLogin(error: error)
+
+        // THEN
+        XCTAssertEqual(sut.appState, .unauthenticated(error: error))
+        XCTAssertTrue(delegate.wasNotified)
+    }
+
     func testThatAppStateChanges_OnSessionLockChange() {
         // GIVEN
         let userSession = UserSessionMock()
