+++ wire-ios/Wire-iOS/Sources/Authentication/AuthenticationFlowStep.swift
@@ -67,7 +67,7 @@ indirect enum AuthenticationFlowStep: Equatable {
 
     // Post Sign-In
     case noHistory(credentials: ZMCredentials?, context: NoHistoryContext)
-    case clientManagement(clients: [UserClient], credentials: ZMCredentials?)
+    case clientManagement(clients: [UserClient])
     case deleteClient(clients: [UserClient])
     case addEmailAndPassword
     case enrollE2EIdentity
