+++ wire-ios/Wire-iOS/Sources/Authentication/Event Handlers/Input/AuthenticationButtonTapInputHandler.swift
@@ -38,8 +38,7 @@ final class AuthenticationButtonTapInputHandler: AuthenticationEventHandler {
             return [.showLoadingView, .startE2EIEnrollment]
         case .noHistory:
             return [.showLoadingView, .configureNotifications, .completeBackupStep]
-            // TODO Katerina remove credentials from clientManagement
-        case .clientManagement(let clients, let credentials):
+        case .clientManagement(let clients):
             let nextStep = AuthenticationFlowStep.deleteClient(clients: clients)
             return [AuthenticationCoordinatorAction.transition(nextStep, mode: .normal)]
         case .pendingEmailLinkVerification:
