+++ wire-ios/Wire-iOS/Sources/Authentication/Event Handlers/Post-Login/Registration Error/AuthenticationClientLimitErrorHandler.swift
@@ -48,33 +48,10 @@ final class AuthenticationClientLimitErrorHandler: AuthenticationEventHandler {
             return nil
         }
 
-        guard let nextStep = AuthenticationFlowStep.makeClientManagementStep(from: error, credentials: authenticationCredentials, statusProvider: self.statusProvider) else {
+        guard let nextStep = AuthenticationFlowStep.makeClientManagementStep(from: error, statusProvider: self.statusProvider) else {
             return nil
         }
 
-        return [.hideLoadingView, .transition(nextStep, mode: .reset)]
-    }
-
-}
-
-final class AuthenticationClientLimitErrorHandler1: AuthenticationEventHandler {
-
-    weak var statusProvider: AuthenticationStatusProvider?
-
-    func handleEvent(currentStep: AuthenticationFlowStep, context: (NSError?, Int)) -> [AuthenticationCoordinatorAction]? {
-        let (error, _) = context
-
-        // Only handle canNotRegisterMoreClients errors
-        guard
-            let error = error,
-            error.userSessionErrorCode == .canNotRegisterMoreClients,
-            let nextStep = AuthenticationFlowStep.makeClientManagementStep(
-                from: error,
-                credentials: nil,
-                statusProvider: self.statusProvider)
-        else {
-            return nil
-        }
         return [.hideLoadingView, .transition(nextStep, mode: .reset)]
     }
